Move utilities classes to package utils
