+++ code/larin-anton/ViewModel/src/main/java/ru/unn/agile/PersonalFinance/ViewModel/utils/StringHelper.java
@@ -1,32 +1,30 @@
-package ru.unn.agile.PersonalFinance.ViewModel;
+package ru.unn.agile.PersonalFinance.ViewModel.utils;
 
 import javafx.beans.binding.Bindings;
 import javafx.beans.binding.BooleanBinding;
-import javafx.beans.binding.StringBinding;
-import javafx.beans.property.BooleanProperty;
 import javafx.beans.property.StringProperty;
 
-final class StringHelper {
+public final class StringHelper {
     static boolean isNullOrEmpty(final String value) {
         return value == null ||
                value.trim().isEmpty();
     }
 
-    static boolean isEqual(final String lhs, final String rhs) {
+    public static boolean isEqual(final String lhs, final String rhs) {
         if (lhs != null && rhs != null) {
             return lhs.equals(rhs);
         }
         return lhs == null && rhs == null;
     }
 
-    static boolean areEqualTrimmed(final String lhs, final String rhs) {
+    public static boolean areEqualTrimmed(final String lhs, final String rhs) {
         if (lhs != null && rhs != null) {
             return lhs.trim().equals(rhs.trim());
         }
         return lhs == null && rhs == null;
     }
 
-    static BooleanBinding isNullOrEmpty(final StringProperty property) {
+    public static BooleanBinding isNullOrEmpty(final StringProperty property) {
         return Bindings.createBooleanBinding(() ->
                 StringHelper.isNullOrEmpty(property.get()), property);
     }
